+++ core/api_container/api/kurtosis_service.go
@@ -9,20 +9,18 @@ import (
 	"context"
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/kurtosis/api_container/execution/test_execution_status"
-	"github.com/kurtosis-tech/kurtosis/commons"
+	"github.com/kurtosis-tech/kurtosis/api_container/service_network"
+	"github.com/kurtosis-tech/kurtosis/api_container/service_network/partition_topology"
+	"github.com/kurtosis-tech/kurtosis/api_container/service_network/topology_types"
+	"github.com/kurtosis-tech/kurtosis/commons/docker_manager"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
-	"net"
 	"net/http"
 	"strings"
-	"sync"
 	"time"
 )
 
 const (
-
-	// How long we'll wait when making a best-effort attempt to stop a container
-	containerStopTimeout = 15 * time.Second
 
 	// Prefixes for the two types of threads we'll spin up
 	completionTimeoutPrefix = "Test completion/timeout thread"
@@ -35,44 +33,30 @@ type KurtosisService struct {
 	//  exactly once
 	testSuiteContainerId string
 
-	dockerManager *commons.DockerManager
+	dockerManager *docker_manager.DockerManager
 
-	dockerNetworkId string
-
-	freeIpAddrTracker *commons.FreeIpAddrTracker
+	serviceNetwork *service_network.ServiceNetwork
 
 	// A value will be pushed to this channel when the status of the execution of a test changes, e.g. via the test suite
 	//  registering that execution has started, or the timeout has been hit, etc.
 	testExecutionStatusChan chan test_execution_status.TestExecutionStatus
 
-	// The name of the Docker volume for this test that will be mounted on all services
-	testVolumeName string
-
-	// The names of the tests inside the suite; will be nil if no test suite has been registered yet
-	suiteTestNames []string
-
 	// Flag that will only be switched to true once, indicating that a test execution has been registered
 	testExecutionRegistered bool
-
-	mutex *sync.Mutex
 }
 
 func NewKurtosisService(
 		testSuiteContainerId string,
 		testExecutionStatusChan chan test_execution_status.TestExecutionStatus,
-		dockerManager *commons.DockerManager,
-		dockerNetworkId string,
-		freeIpAddrTracker *commons.FreeIpAddrTracker,
-		testVolumeName string) *KurtosisService {
+		dockerManager *docker_manager.DockerManager,
+		serviceNetwork *service_network.ServiceNetwork) *KurtosisService {
+
 	return &KurtosisService{
 		testSuiteContainerId:    testSuiteContainerId,
-		testExecutionStatusChan: testExecutionStatusChan,
 		dockerManager:           dockerManager,
-		dockerNetworkId:         dockerNetworkId,
-		freeIpAddrTracker:       freeIpAddrTracker,
-		testVolumeName: testVolumeName,
-		suiteTestNames:          nil,
-		mutex:                   &sync.Mutex{},
+		serviceNetwork:          serviceNetwork,
+		testExecutionStatusChan: testExecutionStatusChan,
+		testExecutionRegistered: false,
 	}
 }
 
@@ -80,10 +64,25 @@ func NewKurtosisService(
 Adds a service with the given parameters to the network
  */
 func (service *KurtosisService) AddService(httpReq *http.Request, args *AddServiceArgs, result *AddServiceResponse) error {
-	service.mutex.Lock()
-	defer service.mutex.Unlock()
-
 	logrus.Infof("Received request to add a service with the following args: %v", *args)
+
+	serviceIdStr := strings.TrimSpace(args.ServiceID)
+	if serviceIdStr == "" {
+		return stacktrace.NewError("Service ID cannot be empty or whitespace")
+	}
+	serviceId := topology_types.ServiceID(serviceIdStr)
+
+	partitionId := topology_types.PartitionID(args.PartitionID)
+
+	imageNameStr := strings.TrimSpace(args.ImageName)
+	if imageNameStr == "" {
+		return stacktrace.NewError("Image name cannot be empty or whitespace")
+	}
+
+	ipPlaceholderStr := strings.TrimSpace(args.IPPlaceholder)
+	if ipPlaceholderStr == "" {
+		return stacktrace.NewError("IP placeholder string cannot be empty or whitespace")
+	}
 
 	usedPorts := map[nat.Port]bool{}
 	for _, portSpecStr := range args.UsedPorts {
@@ -105,79 +104,104 @@ func (service *KurtosisService) AddService(httpReq *http.Request, args *AddServi
 		usedPorts[portObj] = true
 	}
 
-	freeIp, err := service.freeIpAddrTracker.GetFreeIpAddr()
-	if err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred when getting an IP to give the container running the new service with Docker image '%v'",
-			args.ImageName)
-	}
-	logrus.Debugf("Giving new service the following IP: %v", freeIp.String())
-
-	// The user won't know the IP address, so we'll need to replace all the IP address placeholders with the actual
-	//  IP
-	replacedStartCmd, replacedEnvVars := replaceIpPlaceholderForDockerParams(
-		args.IPPlaceholder,
-		freeIp,
-		args.StartCmd,
-		args.DockerEnvironmentVars)
-
-	portBindings := map[nat.Port]*nat.PortBinding{}
-	for port, _ := range usedPorts {
-		portBindings[port] = nil
-	}
-
-	containerId, err := service.dockerManager.CreateAndStartContainer(
+	serviceIp, err := service.serviceNetwork.AddServiceInPartition(
 		httpReq.Context(),
-		args.ImageName,
-		service.dockerNetworkId,
-		freeIp,
-		portBindings,
-		replacedStartCmd,
-		replacedEnvVars,
-		map[string]string{}, // no bind mounts for services created via the Kurtosis API
-		map[string]string{
-			service.testVolumeName: args.TestVolumeMountFilepath,
-		},
-	)
+		serviceId,
+		imageNameStr,
+		usedPorts,
+		partitionId,
+		ipPlaceholderStr,
+		args.StartCmd,
+		args.DockerEnvironmentVars,
+		args.TestVolumeMountDirpath)
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred starting the Docker container for the new service")
+		return stacktrace.Propagate(err, "An error occurred creating service '%v' inside partition '%v'", serviceId, partitionId)
 	}
+	logrus.Infof("Successfully added service '%v'", serviceId)
 
-	result.IPAddress = freeIp.String()
-	result.ContainerID = containerId
-
-	logrus.Infof("Successfully added service")
+	result.IPAddress = serviceIp.String()
 	return nil
 }
 
 /*
 Removes the service with the given service ID from the network
  */
-func (service *KurtosisService) RemoveService(httpReq *http.Request, args *RemoveServiceArgs, result *interface{}) error {
-	service.mutex.Lock()
-	defer service.mutex.Unlock()
+func (service *KurtosisService) RemoveService(httpReq *http.Request, args *RemoveServiceArgs, _ *interface{}) error {
+	logrus.Infof("Received request to remove a service with the following args: %v", *args)
 
-	containerId := args.ContainerID
-	logrus.Debugf("Removing container ID %v...", containerId)
-
-	// Make a best-effort attempt to stop the container
-	err := service.dockerManager.StopContainer(httpReq.Context(), containerId, containerStopTimeout)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred stopping the container with ID %v", containerId)
+	serviceIdStr := strings.TrimSpace(args.ServiceID)
+	if serviceIdStr == "" {
+		return stacktrace.NewError("Service ID cannot be empty or whitespace")
 	}
-	logrus.Debugf("Successfully removed service with container ID %v", containerId)
+	serviceId := topology_types.ServiceID(serviceIdStr)
+
+	if args.ContainerStopTimeoutSeconds <= 0 {
+		return stacktrace.NewError("Container stop timeout seconds cannot be <= 0")
+	}
+
+	containerStopTimeout := time.Duration(args.ContainerStopTimeoutSeconds) * time.Second
+	if err := service.serviceNetwork.RemoveService(httpReq.Context(), serviceId, containerStopTimeout); err != nil {
+		return stacktrace.Propagate(err, "An error occurred removing service with ID '%v'", serviceId)
+	}
 
 	return nil
 }
 
+func (service *KurtosisService) Repartition(httpReq *http.Request, args *RepartitionArgs, _ *interface{}) error {
+	logrus.Infof("Received request to repartition the test network with the following args: %v", args)
+
+	// No need to check for dupes here - that happens at the lowest-level call to ServiceNetwork.Repartition (as it should)
+	partitionServices := map[topology_types.PartitionID]*topology_types.ServiceIDSet{}
+	for partitionIdStr, serviceIdStrSet := range args.PartitionServices {
+		partitionId := topology_types.PartitionID(partitionIdStr)
+		serviceIdSet := topology_types.NewServiceIDSet()
+		for serviceIdStr := range serviceIdStrSet {
+			serviceId := topology_types.ServiceID(serviceIdStr)
+			serviceIdSet.AddElem(serviceId)
+		}
+		partitionServices[partitionId] = serviceIdSet
+	}
+
+	partitionConnections := map[topology_types.PartitionConnectionID]partition_topology.PartitionConnection{}
+	for partitionAStr, partitionBToConnection := range args.PartitionConnections {
+		partitionAId := topology_types.PartitionID(partitionAStr)
+		for partitionBStr, connectionInfo := range partitionBToConnection {
+			partitionBId := topology_types.PartitionID(partitionBStr)
+			partitionConnectionId := *topology_types.NewPartitionConnectionID(partitionAId, partitionBId)
+			if _, found := partitionConnections[partitionConnectionId]; found {
+				return stacktrace.NewError(
+					"Partition connection '%v' <-> '%v' was defined twice (possibly in reverse order)",
+					partitionAId,
+					partitionBId)
+			}
+			partitionConnection := partition_topology.PartitionConnection{
+				IsBlocked: connectionInfo.IsBlocked,
+			}
+			partitionConnections[partitionConnectionId] = partitionConnection
+		}
+	}
+
+	defaultConnectionInfo := args.DefaultConnection
+	defaultConnection := partition_topology.PartitionConnection{
+		IsBlocked: defaultConnectionInfo.IsBlocked,
+	}
+
+	if err := service.serviceNetwork.Repartition(
+			httpReq.Context(),
+			partitionServices,
+			partitionConnections,
+			defaultConnection); err != nil {
+		return stacktrace.Propagate(err, "An error occurred repartitioning the test network")
+	}
+	return nil
+}
+
+
 // Registers that the test suite container is going to run a test, and the Kurtosis API container should wait for the
 //  given amount of time before calling the test lost
-func (service *KurtosisService) RegisterTestExecution(httpReq *http.Request, args *RegisterTestExecutionArgs, result *struct{}) error {
-	service.mutex.Lock()
-	defer service.mutex.Unlock()
+func (service *KurtosisService) RegisterTestExecution(_ *http.Request, args *RegisterTestExecutionArgs, _ *struct{}) error {
+	logrus.Infof("Received request to register a test execution with timeout of %v seconds", args.TestTimeoutSeconds)
 
-	logrus.Infof("Received request to register a test execution with timeout of %v seconds...", args.TestTimeoutSeconds)
 	if service.testExecutionRegistered {
 		return stacktrace.NewError("A test execution is already registered with the API container")
 	}
@@ -191,27 +215,26 @@ func (service *KurtosisService) RegisterTestExecution(httpReq *http.Request, arg
 	return nil
 }
 
-
 // ============================ Private helper functions ==============================================================
 /*
 Waits for either a) the testsuite container to exit or b) the given timeout to be reached, and pushes a corresponding
 	boolean value to the given channel based on which condition was hit
  */
 func awaitTestCompletionOrTimeout(
-			dockerManager *commons.DockerManager,
+			dockerManager *docker_manager.DockerManager,
 			testSuiteContainerId string,
 			timeoutSeconds int,
 			testExecutionStatusChan chan test_execution_status.TestExecutionStatus) {
 	logrus.Debugf("[%v] Thread started", completionTimeoutPrefix)
 
-	context, cancelFunc := context.WithCancel(context.Background())
+	cancellableCtx, cancelFunc := context.WithCancel(context.Background())
 	defer cancelFunc()
 
 	// Kick off a thread that will only exit upon a) the testsuite container exiting or b) the context getting cancelled
 	testSuiteContainerExitedChan := make(chan struct{})
 
 	logrus.Debugf("[%v] Launching thread to await testsuite container exit...", completionTimeoutPrefix)
-	go awaitTestSuiteContainerExit(context, dockerManager, testSuiteContainerId, testSuiteContainerExitedChan)
+	go awaitTestSuiteContainerExit(cancellableCtx, dockerManager, testSuiteContainerId, testSuiteContainerExitedChan)
 	logrus.Debugf("[%v] Launched thread to await testsuite container exit", completionTimeoutPrefix)
 
 	logrus.Debugf("[%v] Blocking until either the test suite exits or the test timeout is hit...", completionTimeoutPrefix)
@@ -235,7 +258,7 @@ Waits for the container to exit until the context is cancelled
  */
 func awaitTestSuiteContainerExit(
 		context context.Context,
-		dockerManager *commons.DockerManager,
+		dockerManager *docker_manager.DockerManager,
 		testSuiteContainerId string,
 		testSuiteContainerExitedChan chan struct{}) {
 	logrus.Debugf("[%v] Thread started", completionPrefix)
@@ -253,27 +276,4 @@ func awaitTestSuiteContainerExit(
 	//  after the timeout, this won't do anything because nobody will be monitoring the other end of the channel
 	close(testSuiteContainerExitedChan)
 	logrus.Debugf("[%v] Thread is exiting", completionPrefix)
-}
-
-/*
-Small helper function to replace the IP placeholder with the real IP string in the start command and Docker environment
-	variables.
- */
-func replaceIpPlaceholderForDockerParams(
-		ipPlaceholder string,
-		realIp net.IP,
-		startCmd []string,
-		envVars map[string]string) ([]string, map[string]string) {
-	ipPlaceholderStr := ipPlaceholder
-	replacedStartCmd := []string{}
-	for _, cmdFragment := range startCmd {
-		replacedCmdFragment := strings.ReplaceAll(cmdFragment, ipPlaceholderStr, realIp.String())
-		replacedStartCmd = append(replacedStartCmd, replacedCmdFragment)
-	}
-	replacedEnvVars := map[string]string{}
-	for key, value := range envVars {
-		replacedValue := strings.ReplaceAll(value, ipPlaceholderStr, realIp.String())
-		replacedEnvVars[key] = replacedValue
-	}
-	return replacedStartCmd, replacedEnvVars
 }