+++ core/commons/docker_manager/container_capabilities.go
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2020 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package docker_manager
+
+const (
+	NetAdmin ContainerCapability = "NET_ADMIN"
+)
+
+type ContainerCapability string